web/requirements.txt
Django==1.8.1
gunicorn==19.3.0
psycopg2==2.6
redis==2.10.3
django-honeypot==0.5.0